+++ src/main/java/net/geforcemods/securitycraft/blocks/SonicSecuritySystemBlock.java
@@ -10,16 +10,12 @@ import net.geforcemods.securitycraft.network.client.OpenScreen.DataType;
 import net.geforcemods.securitycraft.util.LevelUtils;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
-import net.minecraft.core.component.DataComponents;
-import net.minecraft.nbt.CompoundTag;
-import net.minecraft.nbt.Tag;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.ItemInteractionResult;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.component.CustomData;
 import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
@@ -160,14 +156,10 @@ public class SonicSecuritySystemBlock extends OwnableBlock implements SimpleWate
 
 	@Override
 	public ItemStack getCloneItemStack(BlockState state, HitResult target, LevelReader level, BlockPos pos, Player player) {
-		CompoundTag blockTag = level.getBlockEntity(pos).getUpdateTag(level.registryAccess());
-		ItemStack stack = new ItemStack(SCContent.SONIC_SECURITY_SYSTEM_ITEM.get());
+		if (level.getBlockEntity(pos) instanceof SonicSecuritySystemBlockEntity be)
+			return be.getItem();
 
-		if (!blockTag.contains("LinkedBlocks"))
-			return stack;
-
-		CustomData.update(DataComponents.CUSTOM_DATA, stack, tag -> tag.put("LinkedBlocks", blockTag.getList("LinkedBlocks", Tag.TAG_COMPOUND)));
-		return stack;
+		return super.getCloneItemStack(state, target, level, pos, player);
 	}
 
 	@Override