+++ src/main/java/net/geforcemods/securitycraft/datagen/BlockLootTableGenerator.java
@@ -14,7 +14,6 @@ import net.geforcemods.securitycraft.misc.BlockEntityNBTCondition;
 import net.minecraft.advancements.critereon.StatePropertiesPredicate;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.HolderLookup;
-import net.minecraft.core.component.DataComponents;
 import net.minecraft.data.loot.LootTableSubProvider;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.Nameable;
@@ -134,7 +133,7 @@ public class BlockLootTableGenerator implements LootTableSubProvider {
 						.setRolls(ConstantValue.exactly(1))
 						.add(LootItem.lootTableItem(SCContent.SONIC_SECURITY_SYSTEM_ITEM.get())
 								.apply(CopyComponentsFunction.copyComponents(CopyComponentsFunction.Source.BLOCK_ENTITY)
-										.include(DataComponents.CUSTOM_DATA)) //TODO: Check if this works
+										.include(SCContent.INDEXED_POSITIONS.get())) //TODO: Check if this works
 								.apply(CopyNameFunction.copyName(NameSource.BLOCK_ENTITY)))));
 		//@formatter:on