+++ src/main/java/net/geforcemods/securitycraft/network/server/RemovePositionFromSSS.java
@@ -2,9 +2,9 @@ package net.geforcemods.securitycraft.network.server;
 
 import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.SecurityCraft;
-import net.geforcemods.securitycraft.items.SonicSecuritySystemItem;
+import net.geforcemods.securitycraft.components.IndexedPositions;
 import net.geforcemods.securitycraft.util.PlayerUtils;
-import net.minecraft.core.BlockPos;
+import net.minecraft.core.GlobalPos;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
@@ -13,11 +13,11 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.neoforged.neoforge.network.handling.IPayloadContext;
 
-public record RemovePositionFromSSS(BlockPos pos) implements CustomPacketPayload {
+public record RemovePositionFromSSS(GlobalPos globalPos) implements CustomPacketPayload {
 	public static final Type<RemovePositionFromSSS> TYPE = new Type<>(new ResourceLocation(SecurityCraft.MODID, "remove_position_from_sss"));
 	//@formatter:off
 	public static final StreamCodec<RegistryFriendlyByteBuf, RemovePositionFromSSS> STREAM_CODEC = StreamCodec.composite(
-			BlockPos.STREAM_CODEC, RemovePositionFromSSS::pos,
+			GlobalPos.STREAM_CODEC, RemovePositionFromSSS::globalPos,
 			RemovePositionFromSSS::new);
 	//@formatter:on
 
@@ -31,6 +31,6 @@ public record RemovePositionFromSSS(BlockPos pos) implements CustomPacketPayload
 		ItemStack stack = PlayerUtils.getItemStackFromAnyHand(player, SCContent.SONIC_SECURITY_SYSTEM_ITEM.get());
 
 		if (!stack.isEmpty())
-			SonicSecuritySystemItem.removeLinkedBlock(stack, pos);
+			stack.getOrDefault(SCContent.INDEXED_POSITIONS, IndexedPositions.EMPTY).remove(stack, globalPos);
 	}
 }