+++ Changelog.md
@@ -1,4 +1,4 @@
 --------------------------Changelog for v1.9.9-beta1 of SecurityCraft--------------------------
 
 - Change: The "taser_effects" and "powered_taser_effects" config settings have been moved to the item component "minecraft:potion_contents" on the taser
-- Fix: The Mine Remote Access Tool neither saves nor respects the level of the bound mines
+- Fix: The Mine/Sentry Remote Access Tool and Sonic Security System neither saves nor respects the level of the bound objects